fix search
